requirements.txt
apipkg==1.4
attrs==17.4.0
cached-property==1.3.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.6
colorama==0.3.9
coverage==4.4.2
dj-database-url==0.4.2
Django==1.11.8
django-debug-toolbar==1.9.1
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
execnet==1.5.0
flake8==3.5.0
flake8-blind-except==0.1.1
flake8-bugbear==17.12.0
flake8-builtins==1.0.post0
flake8-commas==1.0.0
flake8-comprehensions==1.4.1
flake8-docstrings==1.3.0
flake8-pep3101==1.2
flake8-polyfill==1.0.2
flake8-quotes==0.13.0
gunicorn==19.7.1
hangman==0.0.1
identify==1.0.7
idna==2.6
isort==4.2.15
mccabe==0.6.1
nodeenv==1.2.0
numpy==1.13.3
olefile==0.44
ordereddict==1.1
pep8-naming==0.5.0
pew==1.1.0
Pillow==4.3.0
psutil==5.3.1
py==1.4.34
pycodestyle==2.3.1
pydocstyle==2.1.1
pyflakes==1.6.0
PyJWT==1.5.3
pytest==3.2.3
pytest-cache==1.0
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-flake8==0.9.1
pytest-forked==0.2
pytest-isort==0.1.0
pytest-xdist==1.22.0
pytz==2017.3
requests==2.18.4
six==1.11.0
snowballstemmer==1.2.1
sqlparse==0.2.4
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
whitenoise==3.3.1
